+++ main.go
@@ -25,6 +25,7 @@ import (
 	"time"
 
 	log "github.com/Sirupsen/logrus"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
 
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
@@ -66,7 +67,7 @@ func main() {
 
 	stopChan := make(chan struct{}, 1)
 
-	go registerHandlers(cfg.HealthPort)
+	go serveMetrics(cfg.MetricsAddress)
 	go handleSigterm(stopChan)
 
 	client, err := newClient(cfg)
@@ -115,14 +116,6 @@ func main() {
 	}
 }
 
-func registerHandlers(port string) {
-	http.HandleFunc("/healthz", func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-		w.Write([]byte("OK"))
-	})
-	log.Fatal(http.ListenAndServe(fmt.Sprintf(":%s", port), nil))
-}
-
 func handleSigterm(stopChan chan struct{}) {
 	signals := make(chan os.Signal, 1)
 	signal.Notify(signals, syscall.SIGTERM)
@@ -149,4 +142,15 @@ func newClient(cfg *externaldns.Config) (*kubernetes.Clientset, error) {
 	}
 
 	return client, nil
+}
+
+func serveMetrics(address string) {
+	http.HandleFunc("/healthz", func(w http.ResponseWriter, _ *http.Request) {
+		w.WriteHeader(http.StatusOK)
+		w.Write([]byte("OK"))
+	})
+
+	http.Handle("/metrics", promhttp.Handler())
+
+	log.Fatal(http.ListenAndServe(address, nil))
 }