+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+Features:
+
+  - Support creation of CNAME records when endpoint target is a hostname.
+
 Bug fixes:
 
   - AWS Route 53: Do not submit request when there are no changes.