+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Bug fixes:
 
   - AWS Route 53: Do not submit request when there are no changes.
 
+Documentation:
+
+  - Add documentation on how to setup ExternalDNS for Services on AWS.
+
 ## v0.1.0 - 2017-03-30 (KubeCon)
 
 Features: